# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  jsonreqresp.AddEventRequest:
    properties:
      address:
        example: ул. Пречистенка, 12/2
        maxLength: 500
        type: string
      artworkIDs:
        items:
          type: string
        type: array
      canVisit:
        example: true
        type: boolean
      cntTickets:
        example: 100
        minimum: 0
        type: integer
      dateBegin:
        example: "2023-06-15T10:00:00Z"
        type: string
      dateEnd:
        example: "2023-09-20T18:00:00Z"
        type: string
      title:
        example: Ночная выставка
        maxLength: 255
        type: string
    required:
    - address
    - canVisit
    - cntTickets
    - dateBegin
    - dateEnd
    - title
    type: object
...
  jsonreqresp.ArtworkResponse:
    properties:
      author:
        $ref: '#/definitions/jsonreqresp.AuthorResponse'
      collection:
        $ref: '#/definitions/jsonreqresp.CollectionResponse'
      creationYear:
        example: 1503
        type: integer
      id:
        example: bb2e8400-e29b-41d4-a716-446655442222
        type: string
      material:
        example: Poplar wood
        type: string
      size:
        example: 77 cm × 53 cm
        type: string
      technic:
        example: Oil painting
        type: string
      title:
        example: Mona Lisa
        type: string
    type: object
  jsonreqresp.AuthorResponse:
    properties:
      birthYear:
        example: 1452
        type: integer
      deathYear:
        example: 1519
        type: integer
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: Leonardo da Vinci
        type: string
    type: object
...
  jsonreqresp.CollectionResponse:
    properties:
      id:
        example: aa1e8400-e29b-41d4-a716-446655441111
        type: string
      title:
        example: Louvre Museum Collection
        type: string
    type: object
  jsonreqresp.ConArtworkEventRequest:
    properties:
      artworkID:
        type: string
    required:
    - artworkID
    type: object
...
  jsonreqresp.DeleteEventRequest:
    properties:
      id:
        type: string
    required:
    - id
    type: object
...
  jsonreqresp.EventResponse:
    properties:
      address:
        example: ул. Пречистенка, 12/2
        type: string
      artworkIDs:
        items:
          type: string
        type: array
      canVisit:
        example: true
        type: boolean
      cntTickets:
        example: 150
        type: integer
      dateBegin:
        example: "2023-06-15T10:00:00Z"
        type: string
      dateEnd:
        example: "2023-09-20T18:00:00Z"
        type: string
      employeeID:
        example: cfd9ff5d-cb37-407c-b043-288a482e9239
        type: string
      id:
        example: bb2e8400-e29b-41d4-a716-446655442222
        type: string
      title:
        example: Выставка импрессионистов
        type: string
      valid:
        example: true
        type: boolean
    type: object
...
  jsonreqresp.UpdateEventRequest:
    properties:
      address:
        example: ул. Пречистенка, 12/2
        maxLength: 500
        type: string
      canVisit:
        example: true
        type: boolean
      cntTickets:
        example: 100
        minimum: 0
        type: integer
      dateBegin:
        example: "2023-06-15T10:00:00Z"
        type: string
      dateEnd:
        example: "2023-09-20T18:00:00Z"
        type: string
      id:
        example: 44a315d0-663c-4813-92a6-d7977c2f2aba
        type: string
      title:
        example: Ночная выставка
        maxLength: 255
        type: string
    required:
    - address
    - canVisit
    - cntTickets
    - dateBegin
    - dateEnd
    - id
    - title
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API для системы учета произведений искусств
  title: Музей
  version: "1.0"
paths:
...
  /employee/events:
    delete:
      consumes:
      - application/json
      description: Удаляет существующее мероприятие
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      - description: Данные для удаления мероприятия
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/jsonreqresp.DeleteEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Мероприятие успешно удалено
        "400":
          description: Неверный запрос - ошибка валидации
        "404":
          description: Не найдено - мероприятие не найдено
      security:
      - ApiKeyAuth: []
      summary: Удалить мероприятие (сотрудник)
      tags:
      - Мероприятия
    get:
      description: Возвращает список всех мероприятий
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/jsonreqresp.EventResponse'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Получить все мероприятия (сотрудник)
      tags:
      - Мероприятия
    post:
      consumes:
      - application/json
      description: Создает новое мероприятие
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      - description: Данные мероприятия
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/jsonreqresp.AddEventRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Мероприятие успешно создано
        "400":
          description: Неверный запрос - ошибка валидации
        "401":
          description: Не авторизован
        "404":
          description: Не найдено - сотрудник не найден
      security:
      - ApiKeyAuth: []
      summary: Добавить новое мероприятие (сотрудник)
      tags:
      - Мероприятия
    put:
      consumes:
      - application/json
      description: Обновляет существующее мероприятие
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      - description: Данные для обновления мероприятия
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/jsonreqresp.UpdateEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Мероприятие успешно обновлено
        "400":
          description: Неверный запрос - ошибка валидации
        "404":
          description: Не найдено - мероприятие не найдено
      security:
      - ApiKeyAuth: []
      summary: Обновить мероприятие (сотрудник)
      tags:
      - Мероприятия
  /employee/events/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет произведение из существующего мероприятия
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID мероприятия
        in: path
        name: id
        required: true
        type: string
      - description: Данные для связи произведения с мероприятием
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/jsonreqresp.ConArtworkEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Произведение успешно удалено из мероприятия
        "400":
          description: Неверный запрос - ошибка валидации
        "404":
          description: Не найдено - мероприятие или произведение не найдено
      security:
      - ApiKeyAuth: []
      summary: Удалить произведение из мероприятия (сотрудник)
      tags:
      - Мероприятия
    put:
      consumes:
      - application/json
      description: Добавляет произведение к существующему мероприятию
      parameters:
      - description: Bearer токен
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID мероприятия
        in: path
        name: id
        required: true
        type: string
      - description: Данные для связи произведения с мероприятием
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/jsonreqresp.ConArtworkEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Произведение успешно добавлено к мероприятию
        "400":
          description: Неверный запрос - ошибка валидации или дублирование произведения
        "404":
          description: Не найдено - мероприятие или произведение не найдено
      security:
      - ApiKeyAuth: []
      summary: Добавить произведение к мероприятию (сотрудник)
      tags:
      - Мероприятия
  /employee/events/{id}/artworks:
    get:
      consumes:
      - application/json
      description: Возвращает список всех произведений данного мероприятия
      parameters:
      - description: bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID мероприятия
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/jsonreqresp.ArtworkResponse'
            type: array
        "400":
          description: Неверный запрос - ошибка валидации
        "404":
          description: Не найдено - мероприятие или произведение не найдено
      security:
      - ApiKeyAuth: []
      summary: Получить все произведения мероприятия (сотрудник)
      tags:
      - Мероприятия
...
  /museum/events:
    get:
      consumes:
      - application/json
      description: Возвращает список всех мероприятий с возможностью фильтрации
      parameters:
      - description: Фильтр по названию мероприятия
        in: query
        maxLength: 255
        name: title
        type: string
      - description: 'Фильтр по минимальной дате начала (формат: ГГГГ-ММ-ДД)'
        format: date
        in: query
        name: date_begin
        type: string
      - description: 'Фильтр по максимальной дате окончания (формат: ГГГГ-ММ-ДД)'
        format: date
        in: query
        name: date_end
        type: string
      - description: Фильтр по доступности для посещения
        in: query
        name: can_visit
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/jsonreqresp.EventResponse'
            type: array
        "400":
          description: Неверный формат даты. Используйте ГГГГ-ММ-ДД
      summary: Получить мероприятия
      tags:
      - Поиск
  /museum/events/{id}:
    get:
      consumes:
      - application/json
      description: Возвращает одно мероприятие по его идентификатору
      parameters:
      - description: ID мероприятия
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/jsonreqresp.EventResponse'
        "400":
          description: Неверный формат ID
        "404":
          description: Мероприятие не найдено
      summary: Получить мероприятие по ID
      tags:
      - Поиск
  /museum/events/{id}/artworks:
    get:
      consumes:
      - application/json
      description: Возвращает список всех произведений данного мероприятия
      parameters:
      - description: ID мероприятия
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/jsonreqresp.ArtworkResponse'
            type: array
        "400":
          description: Неверный запрос - ошибка валидации
        "404":
          description: Не найдено - мероприятие или произведение не найдено
      summary: Получить все произведения мероприятия
      tags:
      - Поиск
...
swagger: "2.0"